requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.4
oauthlib==3.3.1
pandas==2.3.3
praw==7.8.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
tweepy==4.16.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.9.0
nltk>=3.8.1
pyarrow>=14.0.0
fastparquet>=2023.10.1
tqdm>=4.66.0
transformers>=4.41.0
torch>=2.2.0
safetensors>=0.4.2